+++ segmentation/segments--krahenbuhl_ECCV_2014_geodesicobjprop/src/eval_size.py
@@ -1,3 +1,4 @@
+# -*- encoding: utf-8
 """
     Copyright (c) 2014, Philipp Krähenbühl
     All rights reserved.
@@ -24,7 +25,7 @@
 	 (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
     SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 """
-from gop import *
+# -*- encoding: utf-8# -*- encoding: utf-8# -*- encoding: utf-8# -*- encoding: utf-8# -*- encoding: utf-8from gop import *
 import numpy as np
 from time import sleep
 from pickle import dump,load
@@ -32,7 +33,7 @@ from util import *
 
 over_segs,segmentations,boxes = loadVOCAndOverSeg( "test", detector='mssf', year="2012" )
 has_box = [len(b)>0 for b in boxes]
-boxes = [np.vstack(b).astype(np.float32) if len(b)>0 else np.zeros((0,4),dtype=np.float32) for b in boxes]
+boxes = [np.vstack(b).astype(np.int32) if len(b)>0 else np.zeros((0,4),dtype=np.int32) for b in boxes]
 
 prop_settings = setupBaseline( 130, 5, 0.8 )
 bo,b_bo,pool_s,box_pool_s = dataset.proposeAndEvaluate( over_segs, segmentations, boxes, proposals.Proposal( prop_settings ) )
